test(cli): Fix failing test

+++ packages/shadcn/src/utils/create-project.ts
@@ -249,9 +249,12 @@ async function createMonorepoProject(
 
     // Write project name to the package.json
     const packageJsonPath = path.join(projectPath, "package.json")
-    const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, "utf8"))
-    packageJson.name = projectPath.split("/").pop()
-    await fs.writeFile(packageJsonPath, JSON.stringify(packageJson, null, 2))
+    if (fs.existsSync(packageJsonPath)) {
+      const packageJsonContent = await fs.readFile(packageJsonPath, "utf8")
+      const packageJson = JSON.parse(packageJsonContent)
+      packageJson.name = projectPath.split("/").pop()
+      await fs.writeFile(packageJsonPath, JSON.stringify(packageJson, null, 2))
+    }
 
     createSpinner?.succeed("Creating a new Next.js monorepo.")
   } catch (error) {
